+++ PaymentGateway.Domain/TestBank/TestBankPaymentRequest.cs
@@ -2,13 +2,13 @@ namespace PaymentGateway.Domain.TestBank
 {
 	public class TestBankPaymentRequest
 	{
-		public decimal AmountInCents { get; }
+		public int AmountInCents { get; }
 		public string CardNumber { get; }
 		public string Expiry { get; }
 		public string CVV { get; }
 		public string Reference { get; }
 
-		public TestBankPaymentRequest(decimal amountInCents, string cardNumber, string expiry, string cvv, string reference)
+		public TestBankPaymentRequest(int amountInCents, string cardNumber, string expiry, string cvv, string reference)
 		{
 			AmountInCents = amountInCents;
 			CardNumber = cardNumber;
